app/requirements.txt
mysql-connector-python
werkzeug
Pillow
